Time: 23 ms (51.05%), Space: 17.6 MB (18.85%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -14,7 +14,7 @@ public:
                 int right = n-1;
                 while(left <right)
                 {
-                    long long sum = nums[i]+nums[j]+nums[left]+nums[right];
+                    long long sum = (long long)nums[i]+nums[j]+nums[left]+nums[right];
                     if( sum >target)
                     {
                         right--;
